Change train helper initialize, add setter/getter for sample prob
before the initial input was always taken from the input sequence (if
provided), even if sample_probability is not non-zero.
this was no problem for the lstm decoder, because the input sequence was
always zero padded, but for the upcoming hierarchical decoder a non zero
input could be given to the bottom decoder.